+++ drizzle/0023_typical_human_robot.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `playlists` ADD `channel_name_template` text DEFAULT '{name}' NOT NULL;